linux-mount-home, script uploded
The Linux Mount Home script automates the process of mounting a secondary drive as the /home directory in Linux. This script simplifies the task of expanding available storage for user home directories by allowing users to designate a secondary drive for home directory storage. It guides users through the process of selecting the drive and specifying the username for the home directory, making it easy to expand storage capacity and manage user data in Linux environments.

Features:

Automates the process of mounting a secondary drive as /home in Linux.
User-friendly interface with step-by-step instructions.
Allows users to specify the secondary drive and username for the home directory.
Improves storage management and scalability for user data in Linux environments.